Backend/requirements.txt
# -------------------------
# Core FastAPI stack
# -------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1

# -------------------------
# Environment variables
# -------------------------
python-dotenv==1.0.1
# -------------------------
# Database (PostgreSQL)
# -------------------------
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# -------------------------
# Data validation
# -------------------------
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0

# -------------------------
# File uploads & images
# -------------------------
python-multipart==0.0.9
pillow==10.2.0

# -------------------------
# PDF → Image conversion
# -------------------------
pdf2image==1.17.0

# -------------------------
# HTTP & async utilities
# -------------------------
httpx==0.27.0

# -------------------------
# AI / ML (future-ready)
# -------------------------
numpy==1.26.4
opencv-python==4.9.0.80
ultralytics==8.1.18
torch>=1.9.0
torchvision>=0.10.0

# -------------------------
# Optional (logging & utils)
# -------------------------
loguru==0.7.2

cloudinary ==1.32.0